+++ spec/rubocop/cop/rspec/return_from_stub_spec.rb
@@ -115,6 +115,31 @@ RSpec.describe RuboCop::Cop::RSpec::ReturnFromStub, :config do
         end
       RUBY
     end
+
+    include_examples 'autocorrect',
+                     'allow(Foo).to receive(:bar) { 42 }',
+                     'allow(Foo).to receive(:bar).and_return(42)'
+
+    include_examples 'autocorrect',
+                     'allow(Foo).to receive(:bar) { { foo: 42 } }',
+                     'allow(Foo).to receive(:bar).and_return({ foo: 42 })'
+
+    include_examples 'autocorrect',
+                     'allow(Foo).to receive(:bar) {}',
+                     'allow(Foo).to receive(:bar).and_return(nil)'
+
+    original = <<-RUBY
+      allow(Foo).to receive(:bar) do
+        'You called ' \\
+        'me'
+      end
+    RUBY
+    corrected = <<-RUBY
+      allow(Foo).to receive(:bar).and_return('You called ' \\
+        'me')
+    RUBY
+
+    include_examples 'autocorrect', original, corrected
   end
 
   context 'with EnforcedStyle `block`' do
@@ -153,5 +178,45 @@ RSpec.describe RuboCop::Cop::RSpec::ReturnFromStub, :config do
         end
       RUBY
     end
+
+    include_examples 'autocorrect',
+                     'allow(Foo).to receive(:bar).and_return(42)',
+                     'allow(Foo).to receive(:bar) { 42 }'
+
+    include_examples 'autocorrect',
+                     'allow(Foo).to receive(:bar).and_return({ foo: 42 })',
+                     'allow(Foo).to receive(:bar) { { foo: 42 } }'
+
+    include_examples 'autocorrect',
+                     'allow(Foo).to receive(:bar).and_return(foo: 42)',
+                     'allow(Foo).to receive(:bar) { { foo: 42 } }'
+
+    original = <<-RUBY
+      allow(Foo).to receive(:bar).and_return(
+        a: 42,
+        b: 43
+      )
+    RUBY
+    corrected = <<-RUBY # Not perfect, but good enough.
+      allow(Foo).to receive(:bar) { { a: 42,
+        b: 43 } }
+    RUBY
+
+    include_examples 'autocorrect', original, corrected
+
+    include_examples 'autocorrect',
+                     'allow(Foo).to receive(:bar).and_return(nil)',
+                     'allow(Foo).to receive(:bar) { nil }'
+
+    original = <<-RUBY
+      allow(Foo).to receive(:bar).and_return('You called ' \\
+        'me')
+    RUBY
+    corrected = <<-RUBY
+      allow(Foo).to receive(:bar) { 'You called ' \\
+        'me' }
+    RUBY
+
+    include_examples 'autocorrect', original, corrected
   end
 end