+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Add `RSpec/ExpectChange` cop to enforce consistent usage of the change matcher. ([@Darhazer][])
 * Add autocorrect support to `RSpec/LetBeforeExamples`. ([@Darhazer][])
 * Fix `RSpec/InstanceVariable` flagging instance variables inside dynamically defined class. ([@Darhazer][])
+* Add autocorrect support for `RSpec/ReturnFromStub` cop. ([@bquorning][])
 
 ## 1.21.0 (2017-12-13)